Move UI-related data into the UI folder
