+++ GalileoFront/app/src/main/java/com/example/galileomastermindboilerplate/ui/LandingPage.java
@@ -1,33 +1,33 @@
-package com.example.galileomastermindboilerplate;
+package com.example.galileomastermindboilerplate.ui;
 
-import android.content.Context;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
+import com.example.galileomastermindboilerplate.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link SatelliteListPage#newInstance} factory method to
+ * Use the {@link LandingPage#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class SatelliteListPage extends Fragment {
+public class LandingPage extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+    private static final String ARG_PARAM1 = "param1";
+    private static final String ARG_PARAM2 = "param2";
 
-    private RecyclerViewAdapter mAdapter;
+    // TODO: Rename and change types of parameters
+    private String mParam1;
+    private String mParam2;
 
-    public SatelliteListPage() {
+    public LandingPage() {
+        // Required empty public constructor
     }
 
     /**
@@ -36,12 +36,14 @@ public class SatelliteListPage extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment SatelliteListPage.
+     * @return A new instance of fragment LandingPage.
      */
     // TODO: Rename and change types and number of parameters
-    public static SatelliteListPage newInstance(String param1, String param2) {
-        SatelliteListPage fragment = new SatelliteListPage();
+    public static LandingPage newInstance(String param1, String param2) {
+        LandingPage fragment = new LandingPage();
         Bundle args = new Bundle();
+        args.putString(ARG_PARAM1, param1);
+        args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
     }
@@ -49,12 +51,16 @@ public class SatelliteListPage extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        if (getArguments() != null) {
+            mParam1 = getArguments().getString(ARG_PARAM1);
+            mParam2 = getArguments().getString(ARG_PARAM2);
+        }
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_satellite_list_page, container, false);
+        return inflater.inflate(R.layout.fragment_landing_page, container, false);
     }
 }
