+++ GalileoFront/app/src/main/java/com/example/galileomastermindboilerplate/ui/SatelliteListPage.java
@@ -1,4 +1,4 @@
-package com.example.galileomastermindboilerplate;
+package com.example.galileomastermindboilerplate.ui;
 
 import android.os.Bundle;
 
@@ -8,24 +8,21 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.galileomastermindboilerplate.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link GeneralStatsPage#newInstance} factory method to
+ * Use the {@link SatelliteListPage#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class GeneralStatsPage extends Fragment {
+public class SatelliteListPage extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
 
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
+    private RecyclerViewAdapter mAdapter;
 
-    public GeneralStatsPage() {
-        // Required empty public constructor
+    public SatelliteListPage() {
     }
 
     /**
@@ -34,14 +31,12 @@ public class GeneralStatsPage extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment GeneralStatsPage.
+     * @return A new instance of fragment SatelliteListPage.
      */
     // TODO: Rename and change types and number of parameters
-    public static GeneralStatsPage newInstance(String param1, String param2) {
-        GeneralStatsPage fragment = new GeneralStatsPage();
+    public static SatelliteListPage newInstance(String param1, String param2) {
+        SatelliteListPage fragment = new SatelliteListPage();
         Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
     }
@@ -49,16 +44,12 @@ public class GeneralStatsPage extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_general_stats_page, container, false);
+        return inflater.inflate(R.layout.fragment_satellite_list_page, container, false);
     }
 }
