+++ GalileoFront/app/src/main/res/layout/fragment_general_stats_page.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GeneralStatsPage">
+    tools:context=".ui.GeneralStatsPage">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
