+++ GalileoFront/app/src/main/res/layout/fragment_satellite_list_page.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SatelliteListPage">
+    tools:context=".ui.SatelliteListPage">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
